requirements.txt
Pillow>=10.0.0
playwright>=1.40.0
PyYAML>=6.0
requests>=2.28.0
python-dateutil>=2.8.0
flask>=3.0.0
